requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.8.2
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
gunicorn==20.1.0
huggingface-hub==0.11.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
more-itertools==9.0.0
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
prompt-toolkit==3.0.36
pycparser==2.21
PyDocX==0.9.10
pyheif==0.7.1
pypandoc==1.10
pyparsing==3.0.9
PyPDF2==3.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.4
python-socketio==5.7.2
pytube==12.1.0
pytz==2022.6
PyYAML==6.0
rabbitmq==0.2.0
redis==4.4.0
regex==2022.10.31
requests==2.28.1
rq==1.11.1
six==1.16.0
SQLAlchemy==1.4.45
textract==1.6.5
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
